% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranlint-package.R
\docType{package}
\name{cranlint-package}
\alias{cranlint}
\alias{cranlint-package}
\title{cranlint: Static Checks for Common CRAN Submission Mistakes}
\description{
A personal-use package that runs static checks against a package source tree to catch recurring mistakes made before CRAN submission, such as DESCRIPTION formatting issues, fixed random seeds, and leftover debugging statements. Complements tools like 'checkhelper' and 'rcmdcheck' rather than replacing them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/djnavarro/cranlint}
  \item \url{https://cranlint.djnavarro.net/}
  \item Report bugs at \url{https://github.com/djnavarro/cranlint/issues}
}

}
\author{
\strong{Maintainer}: Danielle Navarro \email{djnavarro@protonmail.com} (\href{https://orcid.org/0000-0001-7648-6578}{ORCID})

Authors:
\itemize{
  \item Danielle Navarro \email{djnavarro@protonmail.com} (\href{https://orcid.org/0000-0001-7648-6578}{ORCID})
}

}
\keyword{internal}
