% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-installed-packages.R
\name{cl_check_installed_packages}
\alias{cl_check_installed_packages}
\title{Check for calls to installed.packages()}
\usage{
cl_check_installed_packages(path = ".")
}
\arguments{
\item{path}{Path to the package root. Defaults to the current directory.}
}
\value{
A tibble following the cranlint check-result contract; see
\code{AGENTS.md}.
}
\description{
Flags any call to \code{installed.packages()} in \verb{R/}. Per the CRAN Cookbook,
this function can be slow (it reads several files per installed
package) and shouldn't be used to test whether a package is available;
\code{requireNamespace()}/\code{require()} are the recommended replacements (or
\code{find.package()}/\code{system.file()} to locate one, \code{packageDescription()}
for details of a small number of packages).
}
